Large amount of refactoring of the EMSX components
There are several fixes in here.
The separation between orders and routes has been clarified, and also reduced.  All updates are processed, regardless of being an order or route, and the changes are applied onto a BloombergOrder.
BloombergOrder is now based on Name, and Element from Bloomberg's own API.  This provides type-checking, rather than interpretting the field from a string.